backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "ddgs>=9.10.0",
    "docker>=7.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.129.0",
    "geopy>=2.4.1",
    "google-genai>=1.63.0",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "openai>=2.21.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.13.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.33",
    "timezonefinder>=8.2.1",
    "uvicorn[standard]>=0.40.0",
]
